Add circuits::basic that contains basic circuit constructions (#19)
In this stack I'm going to add a couple of "Hello World" circuits as I learn the circuit programming model.

There are some useful common constructions in base64 that I'd like to reuse so I've extracted them into a separate module.

I'm not sure about the name or where it should live in the tree. A more descriptive name like `common_constructions` seems too long, I couldn't think of anything better so I'm open to suggestions.