Add a circuit that tests whether two inputs are equal (#20)
This is a "Hello World" circuit that tests whether the values of two input wires are equal.

I wrote this while trying to understand how the library works (including writing tests), so it might be a useful circuit for people to look at when they first look at the library. We don't have to merge it though if you think it's too simple / redundant.