Add a circuit that verifies hex encoding (#22)
This circuit is similar to the base64 verification circuit, just a bit simpler. Perhaps it's easier to understand for someone looking at the library for the first time.

In any case, I wrote it as an exercise to understand the circuit programming model. We don't have to merge this if you think it's too simple / redundant.

I wasn't sure if it's correct to add the following check as an assertion, or to add it as a check in the circuit.

```
        debug_assert_eq!(
            EN,
            2 * N,
            "HexEncode: encode array length EN must equal 2*N"
        );

```