Add IaddCinCout gate (#76)
This PR adds an IaddCinCout gate for performing 64bit unsigned addition with carry propagation.

This is similar to Iadd32 but also supports cin (carry_in) and cout (carry_out)  wires that will be used for performing a chain of additions to compute the sum of two big integers consisting of several unsigned 64bit limbs.