[frontend] Fix constraints in IcmpUlt (#80)
I added constraint_verification tests for IcmpUlt and these failed. So this PR is an attempt to fix the issue.

The test failed with inputs a=0 and a=1 in the full-adder borrow relation. I am not 100% sure what the issue is with this check but I think it has something to do with hardcoding the carry_in wire to 1 in the constraint.

With this change I do the computation of the `a - b = a + (~b) + 1` using the same approach as the `IaddCinCout` gate which has the correct borrow (or equivalently carry propagation check). This makes the constraint verification tests pass.

Unfortunately, this adds an extra AND constraint to this gate as we need  not_b as a wire to and so the relation `not_b = b ^ all_1` needs to be checked. Perhaps we can have a variation of the `IaddCinCout` circuit constraints where this can be improved.