[frontend] time failed witness fill (#121)
This is useful to measure witness filling performance even when it's not really
populated with any useful values. This works because the computation is more or
less the same. The actual difference is that we might spend some time on
allocating vectors and formatting when filling out the assertion errors.