[frontend] Add bignum module and add circuit (#114)
This PR sets up the bignum module and representation as a slice / vector of wires. These circuits are used in the RSA verification circuit.

To start, I've added a (non-overflowing) bignum add circuit. This will be used internally in the mod/reduce circuit that's added higher up in the stack.