[frontend] Add RS256 verify circuit (#182)
This PR adds a circuit to verify signatures in JWT that use the "alg" header "RS256". This corresponds to the signature verification algorithm RSASSA-PKCS1-v1_5 using SHA-256 as the hash.

The circuit checks that:

1. The PKCS1-v1.5 encoded message (EM) is equal to signature^65537 mod modulus
2. The prefix in the EM is equal to the expected prefix ( DigestInfo, padding etc.) for the RSASSA-PKCS1-v1_5 + SHA256 scheme
3. The hash in the suffix of the EM is equal to the SHA256 hash of the plaintext message.

Tests are added that use the `rsa` library to generate signatures.

`num-bigint` and `num-integer` are added as dependencies because they are used in in `RsaIntermediates` helper struct to populate intermediate wires for the ModReduce circuits used internally in verification.