[mlecheck] Introduce MleCheckProver trait and MleToSumCheckAdaptor (#266)
This is the first in a sequence of PRs that separates the MLE-check protocol from sumcheck. There is a new `MleCheckProver` trait to facilitate this and an decorator that turns and MleCheckProver into a SumcheckProver efficiently.

Why? This is a cleaner separation of concerns and there are also times when we want to invoke the specialized MLE-check verifier.