[frontend] Update Rs256Verifier API to use FixedByteVec for signature / modulus (#281)
When integrating with the zklogin circuit the JWT signature will be
provided to Rs256Verifier as a FixedByteVec (via the Base64 circuit). So
it's more convenient for the API of the Rs256Verifier to accept
FixedByteVec instead of BigNums.

The byte ordering and packing layout of the signature FixedByteVec is
determined by:

1. The byte ordering of the signature in the JWT (big-endian)
2. how the base64 encoded bytes are packed into wires (little-endian)

The Rs256Verifier uses BigNum internally to verify the RSA computations.
So the byte ordering / packing layout of the signature input must be
converted in the circuit to match the format of BigNum. This happens on
the rs256::fixedbytevec_be_to_bignum function.

I've also added FixedByteVec::populate_bytes to pack bytes into wires in
the same way as happens in Base64UrlSafe:{populate_encoded,
populate_decoded}. This is used in the rs256 tests to verify that
population of the Rs256Verifier circuit will work when integrated with
zklogin.

For consistency I've changed the modulus component of Rs256Verifier to
also be a FixedByteVec. This will also be populated from big-endian
bytes in the zklogin circuit.