[frontend] Rs256Verifier: Encapsulate internal RSA computations (#282)
This PR is intended to simplify use of the Rs256Verifier circuit within
the zklogin circuit.

ModReduce circuits to verify the RSA signature verification within the
Rs256Verify circuit. Previously users of the circuit had to provide
wires for these intermediate values in the circuit constructor and then
populate them externally.

When Rs256Verify is used in the zklogin circuit the signature and
modulus values are extracted from the JWT. The intermediate ModReduce
values can be computed from the signature and modulus.

In this PR the wires for intermediate values are added as witnesses
within the Rs256Verify circuit constructor.

A single method Rs256Verify::populate_rsa is then used to populate the
signature, modulus and computation / population of intermediate values
all at once.