[frontend] Add base64 check for zklogin nonce (#311)
The zklogin nonce is contained in the JWT paylaod as a base64 encoded
string. We must base64-decode this string to the bytes of a SHA256 hash
before asserting it is equal to SHA256(concat(vk_u, T_max, r)).

This is expected to increase the total number of constraints so the
snapshot is regenerated.