[frontend] ceck
Introduce a new tool called ceck.

ceck stands for Constraint Equivalence cheCKer.

This tool was created for Claude to lean on during search for optimizations. The
results are inconclusive, but I decided to submit it anyway because this tool
could've been useful a couple of times before already.

I foresee it could evolve into an umbrella tool similiar to clifutil in the
future. For now, it's just an equivalence checker.