[frontend] circuit builder shared ref (#290)
Initially, the builder was supposed to be passed as a mut, or exclusive,
reference. This is actually more annoying that it has to be, so the
whole thing was changed to have interior mutability which allowed us to
switch to shared reference which greatly reduced friction of using the
circuit builder.

This changeset removes the last remnants of the prior approach.