[frontend] Support base64 verification without padding
The header, payload, and signature components of a JWT are encoded
without the base64 padding characters[0]. So we need a version of the
Base64UrlSafe circuit that validates non-padded base64 encoded strings.

For zklogin we still need to support standard padded base64 encoded
strings because the nonce claim in the JWT payload is a base64 encoded
SHA256 hash.

[0] https://datatracker.ietf.org/doc/html/rfc7515#appendix-C