[ci] relax wasm check (#334)
Now we are building absolutely everything under wasm32-wasi which is a bit too
aggressive because not every crate is supposed to be run under wasm.

Specifically,

1. circuits are meant to be compiled offline, ie. on a developer
machine, before being deployed. While for now binius-frontend is a part of the
compilation tree for binius-prover and binius-verifier it's not always going to
be.

2. there is already one developer tool, `zkl`, which is not meant to be used
   under wasm and probably will never be, and a second tool is upcoming.

It seems to me that what we actually care about is making sure that
binius-prover and binius-verifier (and obviously all the upstream crates) are
compiling and this is what this changeset aims to do.