[core] binius-core
Introduce binius-core.

binius-core is a library that is common to the prover and verifier. This would
be the home for constraint system, because the frontend code is not a suitable
place for it.

The frontend right now contains: the definition of the constraint system, the
compiler, the circuits that use the compiler, and the developer tools like ceck
and zkl.

However, all that stuff is not really needed for the prover and verifier. They
only require the constraint system.

Another reason for extracting this code is to designate a place to place common
definitions that the frontend must share with backend. This will become more
evident in the next PR.