[pcs] Fix for PCS w/ Non Trivial Packing Width (#404)
This PR contains a fix for the one bit PCS when P is specified as a PackedBinaryGhash2x128b.

The source of the bug appears to have been both the test setup and the handling of the provers proof initialization logic regarding partial evals of the small field multilinear.

Interestingly, while the fix worked on PackedBinaryGHash2x128b, it does not on PackedBinaryGhash4x128b.

There is still some investigation to be done regarding this.

Current Status:

The current issue is the PackedBinaryGhash4x128b is producing invalid proofs. It is unclear exactly what is causing this, however, it does appear that the evaluation claim is correct. Replacing the generic with the underyling scalar or the 2x128 packed ghash does not affect the validity of the proof.

 The previous issue was related to not all elements of the packed vector being incorporated into the initialize_proof() method of the one bit PCS. The most recent fix addressed this issue, but interestingly, this change did not generalize to large packed fields.