New NTT implementation (#451)
The new NTT follows depth-first computation (instead of a breadth-first computation as previously.)

It now includes tests which check against the evaluation of the respective polynomial at the respective points.

It exposes a `DomainContext` trait which gives access to twiddles. (Essentially the same as what used to be `Vec<TwiddleAccess>>`. Right now, FRI and PCS still use NTT objects allover. The goal is to change this so that FRI and PCS only ever use `DomainContext`. The only place which should actually use an NTT object is `ReedSolomonCode::encode`.

NTT objects are meant to be a simple wrapper around a `DomainContext` which provide a function to execute the butterfly diagram, using twiddles from `DomainContext`. They should not be aware of the "meaning" of what they are doing. They are only algorithms to execute a butterfly diagram in a performant way. Perhaps, we might now even need NTT _objects_, but just make them functions which take in a `DomainContext` as an argument.

Another thing that I suggest to change in the near future is to use an associated type in the `AdditiveNTT` trait instead of a generic. Because one specific NTT struct will implement `AdditiveNTT` only for a single field. So associated types are the correct thing to use here, not generics. Btw, I feel like this is the case in a lot of our code. I think many generics (and `PhantomData`) can be replaced with associated types (usually it's about a field type). I already applied this to `DomainContext`.

Inverse transform has been removed for now. (Not needed.)

One thing that might need a slight overhaul is the input check (preconditions) and the way we deal with "too many shares" in the multithreaded implementation.